requirements.txt
pandas==1.0.3
numpy==1.18.1
matplotlib==3.1.3
seaborn==0.10.0
Pillow==7.0.0
scipy==v1.4.1